requirements.txt
ollama
fastapi
uvicorn
python-dotenv
requests

